+++ src/Application/Contracts/ContractService.cs
@@ -22,8 +22,8 @@ public class ContractService : IContractService
         _repo = repo;
         _search = search;
         _search.AddModule(new NameSearch());
-        _search.AddModule(new SubstringSearch(contract => contract.Description));
-        _search.AddModule(new SubstringSearch(contract => contract.SupplierDescription));
+        _search.AddModule(new BodyTextSearch(contract => contract.Description));
+        _search.AddModule(new BodyTextSearch(contract => contract.SupplierDescription));
     }
 
     /// <inheritdoc />
